# docs/api/tracker-openapi.yaml
openapi: 3.0.3
info:
  title: Martha Tracker API
  description: |
    Comprehensive REST API for the Martha multi-board tracker system.

    The tracker provides hierarchical issue management across multiple worktrees and boards,
    with support for epics, stories, tasks, and bugs. It includes rich metadata, documentation
    linking, agent integration, and real-time collaboration features.

    ## Hierarchical Structure

    The tracker follows a three-level hierarchy:
    - **Worktrees**: Top-level containers representing code repositories or projects
    - **Boards**: Kanban boards within a worktree for organizing issues
    - **Issues**: Individual work items (epics, stories, tasks, bugs)

    ## API Patterns

    - **Hierarchical Routes**: `/api/tracker/worktrees/{worktreeId}/boards/{boardId}/issues`
    - **Legacy Routes**: `/api/tracker/issues` (backward compatible)
    - **Error Format**: Consistent JSON error responses
    - **Authentication**: Cookie-based session auth (credentials: include)
  version: 1.0.0
  contact:
    name: Martha Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:20000/api/tracker
    description: Local development server
  - url: https://api.martha.dev/api/tracker
    description: Production server
...
components:
  schemas:
...
    Board:
      type: object
      required:
        - id
        - worktree_id
        - name
        - description
        - version
        - columns
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique board identifier (slug format)
          example: phase1-temporal-foundation
        worktree_id:
          type: string
          description: Parent worktree ID
          example: martha-dev-v4
        name:
          type: string
          description: Board name
          example: Phase 1 - Temporal Foundation
        description:
          type: string
          description: Board description
          example: Core temporal workflow implementation
        $schema:
          type: string
          nullable: true
          description: JSON schema URL
        version:
          type: integer
          description: Board version number
          example: 1
        sprint:
          $ref: '#/components/schemas/Sprint'
        columns:
          type: array
          items:
            $ref: '#/components/schemas/BoardColumn'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BoardColumn:
      type: object
      required:
        - id
        - name
        - color
        - issue_ids
      properties:
        id:
          type: string
          description: Column identifier
          example: backlog
        name:
          type: string
          description: Column display name
          example: Backlog
        color:
          type: string
          description: Hex color code
          example: "#6B7280"
        wip_limit:
          type: integer
          nullable: true
          description: Work-in-progress limit
          example: 5
        issue_ids:
          type: array
          items:
            type: string
          description: Ordered list of issue IDs in this column
          example: ["issue-001", "issue-002"]
    Sprint:
      type: object
      nullable: true
      required:
        - id
        - name
        - start_date
        - end_date
      properties:
        id:
          type: string
          example: sprint-2026-01
        name:
          type: string
          example: January 2026 Sprint
        start_date:
          type: string
          format: date
          example: "2026-01-01"
        end_date:
          type: string
          format: date
          example: "2026-01-31"
...
    BoardListResponse:
      type: object
      required:
        - boards
        - count
      properties:
        boards:
          type: array
          items:
            $ref: '#/components/schemas/Board'
        count:
          type: integer
...
paths:
...
  /worktrees/{worktreeId}/boards:
    parameters:
      - name: worktreeId
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Boards
      summary: List boards in worktree
      description: Retrieves all boards in a worktree
      operationId: listBoards
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardListResponse'

    post:
      tags:
        - Boards
      summary: Create board
      description: Creates a new board in the worktree
      operationId: createBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
                - description
              properties:
                id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                columns:
                  type: array
                  items:
                    $ref: '#/components/schemas/BoardColumn'
            example:
              id: new-board
              name: New Board
              description: A new kanban board
              columns:
                - id: backlog
                  name: Backlog
                  color: "#6B7280"
                  wip_limit: null
                  issue_ids: []
                - id: in-progress
                  name: In Progress
                  color: "#3B82F6"
                  wip_limit: 5
                  issue_ids: []
                - id: done
                  name: Done
                  color: "#10B981"
                  wip_limit: null
                  issue_ids: []
      responses:
        '201':
          description: Board created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '400':
          description: Invalid request
        '409':
          description: Board already exists
  /worktrees/{worktreeId}/boards/{boardId}:
    parameters:
      - name: worktreeId
        in: path
        required: true
        schema:
          type: string
      - name: boardId
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Boards
      summary: Get board state
      description: Retrieves complete board state including all columns and issue IDs
      operationId: getBoard
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Board not found

    patch:
      tags:
        - Boards
      summary: Update board
      description: Updates board configuration
      operationId: updateBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                columns:
                  type: array
                  items:
                    $ref: '#/components/schemas/BoardColumn'
      responses:
        '200':
          description: Board updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Board not found

    delete:
      tags:
        - Boards
      summary: Delete board
      description: Deletes a board and all its issues
      operationId: deleteBoard
      responses:
        '204':
          description: Board deleted
        '404':
          description: Board not found

  # Hierarchical Issues API
...
  /board/{boardId}:
    parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Issues (Legacy)
      summary: Get board by ID
      description: Legacy endpoint - retrieves a board by ID
      operationId: getBoardLegacy
      deprecated: true
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
...